+++ ui/src/app/pages/access/access-renewal.spec.tsx
@@ -12,7 +12,7 @@ import {findNodesByExactText, findNodesContainingText, waitOneTickAndUpdate} fro
 import {InstitutionApiStub} from 'testing/stubs/institution-api-stub';
 import {ProfileApiStub, ProfileStubVariables} from 'testing/stubs/profile-api-stub';
 import {accessRenewalModules} from 'app/utils/access-utils';
-import {MILLIS_PER_DAY} from 'app/utils/date';
+import {MILLIS_PER_DAY} from 'app/utils/dates';
 
 const EXPIRY_DAYS = 365
 const nowPlusDays = (days: number) => Date.now() + MILLIS_PER_DAY * days;
